python/requirements.txt
pytablewriter
pyrasgo==0.3.3a3
pyyaml